Refactor of components structure
