--- src/lib/atoms/Expand.svelte
@@ -1,16 +1,0 @@
-<svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-square-plus" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
-  <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-  <rect x="4" y="4" width="16" height="16" rx="2" />
-  <line x1="9" y1="12" x2="15" y2="12" />
-  <line x1="12" y1="9" x2="12" y2="15" />
-</svg>
-
-<style>
-  svg {
-    stroke: rgb(102, 229, 191);
-    transform: translateY(1px);
-    margin-left: 0.25rem;
-  }
-
-  :global(body.expand svg line:last-child) { display: none; }
-</style>
