--- src/lib/atoms/Link.svelte
@@ -1,9 +1,0 @@
-<svg xmlns="http://www.w3.org/2000/svg" width="17" height="17" fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="1.5" viewBox="0 0 24 24" >
-  <path stroke="none" d="M0 0h24v24H0z" />
-  <path d="M10 14a3.5 3.5 0 0 0 5 0l4-4a3.5 3.5 0 0 0-5-5l-.5.5" />
-  <path d="M14 10a3.5 3.5 0 0 0-5 0l-4 4a3.5 3.5 0 0 0 5 5l.5-.5" />
-</svg>
-
-<style>
-  svg { transform: rotate(45deg); }
-</style>
