+++ src/lib/molecules/Semester.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Heading from "$lib/organisms/Heading.svelte";
+  import Heading from "$lib/molecules/Heading.svelte";
   import SprintLink from "$lib/molecules/SprintLink.svelte";
   let { semester } = $props();
 </script>
