+++ src/routes/+page.svelte
@@ -1,6 +1,6 @@
 <script>
   import Semesters from "$lib/organisms/Semesters.svelte";
-  import Program from "$lib/organisms/Program.svelte";
+  import Program from "$lib/molecules/Program.svelte";
   let { data } = $props();
   const { title, subtitle, content, semesters } = data;
 </script>
